cdk/requirements.txt
aws-cdk-lib>=2.190.0
constructs>=10.4.0
